requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
ctransformers
transformers
google-generativeai

-e .